% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_macmon.R
\name{parse_macmon_record}
\alias{parse_macmon_record}
\title{produce a table row from macmon json output}
\usage{
parse_macmon_record(json_str)
}
\arguments{
\item{json_str}{single json doc from macmon}
}
\description{
produce a table row from macmon json output
}
\examples{
mmpath = Sys.getenv("MACMON_PATH")
if (file.exists(mmpath)) {
  tf = tempfile()
  pp = processx::process$new(mmpath, c("pipe"), stdout=tf)
  Sys.sleep(5)
  pp$kill()
  records = readLines(tf)
  df <- purrr::map(records, parse_macmon_record) |> purrr::list_rbind()
  tibble::glimpse(df)
} else message("set MACMON_PATH to demonstrate parse_macmon_records")
}
